Move control code insertion to separate script
